requirements.txt
streamlit==1.28.1
mysql-connector-python==8.1.0
pandas==2.0.3
plotly==5.17.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3